Fixed circular import issues
