+++ api/bonsai_api/dependencies.py
@@ -1,19 +1,16 @@
 """API functions that are used as dependencies."""
 
 import logging
-from typing import Annotated, Any, Generator
-from fastapi import Depends, Request, Security, HTTPException, status
+from typing import Annotated, Generator
+from fastapi import Depends, Request, Security, HTTPException
 from fastapi.security import OAuth2PasswordBearer, SecurityScopes
-from pydantic import BaseModel, Field
 from fastapi.security import SecurityScopes
-from jose import JWTError, jwt
-from bonsai_api.models.auth import TokenData
-from bonsai_api.config import ALGORITHM, USER_ROLES, settings
-
+from bonsai_api.crud.user import get_user_by_token
 from bonsai_api.config import settings
+
 from bonsai_api.db.db import MongoDatabase
-from bonsai_api.models.auth import TokenData
 from bonsai_api.models.user import UserOutputDatabase
+from bonsai_api.models.context import ApiRequestContext
 from bonsai_api.db import Database
 
 from api_client.audit_log import AuditLogClient
@@ -33,13 +30,6 @@ def get_audit_log(request: Request) -> AuditLogClient | None:
     LOG.debug("Audit Log has not been configured")
 
 
-class ApiRequestContext(BaseModel):
-    """The context of a Bonsai API call."""
-
-    actor: Actor
-    metadata: dict[str, Any] = Field(default_factory=dict)
-
-
 def get_database(request: Request) -> Generator[MongoDatabase, None, None]:
     """Get database connection."""
     if not hasattr(request.app.state, "db"):
@@ -47,18 +37,18 @@ def get_database(request: Request) -> Generator[MongoDatabase, None, None]:
     return request.app.state.db
     
 
-async def get_logged_in_user(
+async def get_current_user(
         security_scopes: SecurityScopes,
         token: Annotated[str, Depends(oauth2_scheme)],
         db: Database = Depends(get_database),
         ):
     """Look up logged in user from token."""
 
-    return await get_current_user(security_scopes, token, db)
+    return await get_user_by_token(security_scopes, token, db)
 
 
 async def get_current_active_user(
-    current_user: UserOutputDatabase = Security(get_logged_in_user, scopes=["users:me"]),
+    current_user: UserOutputDatabase = Security(get_current_user, scopes=["users:me"]),
 ) -> UserOutputDatabase | None:
     """Get current active user."""
     # disable API authentication
@@ -86,51 +76,3 @@ async def get_request_context(request: Request, user: UserOutputDatabase | None 
         actor = Actor(type=SourceType.USR, id=actor_id),
         metadata = meta
     )
-
-
-async def get_current_user(
-    security_scopes: SecurityScopes,
-    token: str,
-    db: Database,
-) -> UserOutputDatabase | None:
-    """Get current user."""
-    # check if API authentication is disabled
-    if not settings.api_authentication:
-        return None
-
-    if security_scopes.scopes:
-        authenticate_value = f"Bearer scope={security_scopes.scope_str}"
-    else:
-        authenticate_value = "Bearer"
-
-    credentials_exception = HTTPException(
-        status_code=status.HTTP_401_UNAUTHORIZED,
-        detail="Credentials could not be validated",
-        headers={"WWW-Authenticate": authenticate_value},
-    )
-
-    try:
-        payload = jwt.decode(token, settings.secret_key, algorithms=[ALGORITHM])
-        username: str = payload.get("sub")
-        if username is None:
-            raise credentials_exception
-        token_data = TokenData(username=username)
-    except JWTError as error:
-        raise credentials_exception from error
-
-    try:
-        user: UserOutputDatabase = await get_user(db, username=token_data.username)
-    except EntryNotFound as error:
-        raise credentials_exception from error
-    for scope in security_scopes.scopes:
-        users_all_permissions = {
-            perm for user_role in user.roles for perm in USER_ROLES.get(user_role, [])
-        }
-        if not scope in users_all_permissions:
-            raise HTTPException(
-                status_code=status.HTTP_401_UNAUTHORIZED,
-                detail="Credentials could not be validated",
-                headers={"WWW-Authenticate": authenticate_value},
-            )
-    return user
-
